Addressed the review comments.
